requirements.txt
streamlit>=1.32
google-generativeai>=0.7
openpyxl>=3.1
pdfplumber>=0.11
pdf2image>=1.17
Pillow>=10.0
python-dotenv>=1.0
pytesseract>=0.3.10
pytest>=8.0
